+++ firmware/src/neato_serial.h
@@ -75,7 +75,6 @@ public:
     bool setMotorVacuum(bool on, int speedPercent = 80, std::function<void(bool)> callback = nullptr);
     bool setMotorSideBrush(bool on, int powerMw = 5000, std::function<void(bool)> callback = nullptr);
 
-
     // Set a single robot user setting via "SetUserSettings <key> <value>".
     // Invalidates the user settings cache.
     bool setUserSetting(const String& key, const String& value, std::function<void(bool)> callback = nullptr);